Big reshuffle in src structure
