+++ src/matching/__init__.py
@@ -1,7 +1,6 @@
 """ Top-level imports only. """
 
-from .matching import Matching
-from .player import Hospital, Player, Project
 from .game import Game
-from .solvers import HospitalResident, StableMarriage, StudentAllocation
+from .matching import Matching
+from .player import Player
 from .version import __version__
