+++ src/matching/games/__init__.py
@@ -1,4 +1,4 @@
-""" Make the solvers accessible. """
+""" Make the games and their algorithms accessible. """
 
 from .hospital_resident import HospitalResident, hospital_resident
 from .stable_marriage import StableMarriage, stable_marriage
