+++ src/matching/games/student_allocation.py
@@ -2,22 +2,17 @@
 
 from matching import Game, Matching
 
-from .util import delete_pair, match_pair, unmatch_pair
+from .util import delete_pair, match_pair
 
 
 class StudentAllocation(Game):
     """ A class for solving instances of the Student Allocation problem (SA)
     using an adapted Gale-Shapley algorithm. """
 
-    def __init__(self, students, projects, lecturers):
-
-        for student in students:
-            student.matching = None
-        for player in projects + lecturers:
-            player.matching = []
+    def __init__(self, students, projects, faculty):
 
         self.students = students
         self.projects = projects
-        self.lecturers = lecturers
+        self.faculty = faculty
 
         super().__init__()
