+++ src/matching/games/util.py
