+++ src/matching/players/__init__.py
@@ -1,0 +1,5 @@
+""" Make the player classes more accessible. """
+
+from .faculty import Faculty
+from .hospital import Hospital
+from .project import Project
