+++ tests/unit/test_matching.py
@@ -4,17 +4,9 @@ import pytest
 
 from matching import Matching, Player
 
-suitors = [
-    Player("A"),
-    Player("B"),
-    Player("C"),
-]
+suitors = [Player("A"), Player("B"), Player("C")]
 
-reviewers = [
-    Player(1),
-    Player(2),
-    Player(3),
-]
+reviewers = [Player(1), Player(2), Player(3)]
 
 suitors[0].set_prefs(reviewers)
 suitors[1].set_prefs([reviewers[1], reviewers[0], reviewers[2]])
